backend/requirements.txt
fastapi==0.118.3
uvicorn==0.38.0
python-dotenv==1.2.1
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
psycopg2-binary==2.9.11
google-adk==1.20.0
reportlab==4.4.10
pandas==2.3.3
xgboost==3.2.0
scikit-learn==1.8.0
